# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: NoteHub API
  version: "0.1.0"
  description: |
    NoteHub REST API および WebSocket 仕様。

    ## 認証
    - Access Token: `Authorization: Bearer <token>` ヘッダー（15分有効）
    - Refresh Token: HttpOnly Cookie `notehub_refresh_token`（`/api/auth` パス）
    - CSRF: Cookie `notehub_csrf_token` と Header `X-CSRF-Token` の Double Submit（`POST /api/auth/refresh`, `POST /api/auth/logout`）

    ## レスポンス形式
    成功時は `{ "data": ... }`、失敗時は `{ "error": { "code", "message", "requestId?" } }`。
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://localhost:8082
    description: Local integration test
...
paths:
...
  /api/workspaces/{workspaceId}/members:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [Member]
      summary: メンバー一覧
      operationId: listMembers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
    post:
      tags: [Member]
      summary: メンバー追加（ホストのみ）
      operationId: addMember
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AddMemberResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/workspaces/{workspaceId}/members/{userId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/UserId"
    delete:
      tags: [Member]
      summary: メンバー削除（ホストのみ）
      operationId: removeMember
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        additionalProperties: false
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/workspaces/{workspaceId}/users/search:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - name: email
        in: query
        required: true
        schema:
          type: string
          format: email
    get:
      tags: [Member]
      summary: ユーザー検索（ホストのみ）
      operationId: searchUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SearchUserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/documents/{documentId}/versions:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
    get:
      tags: [Version]
      summary: 編集履歴一覧
      operationId: listVersions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/VersionListItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Version]
      summary: 手動バージョン保存
      operationId: saveManualVersion
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SaveManualVersionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/documents/{documentId}/versions/{versionId}:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
      - $ref: "#/components/parameters/VersionId"
    get:
      tags: [Version]
      summary: 編集履歴詳細
      operationId: getVersion
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/GetVersionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/documents/{documentId}/versions/{versionId}/restore:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
      - $ref: "#/components/parameters/VersionId"
    post:
      tags: [Version]
      summary: バージョン復元
      operationId: restoreVersion
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RestoreVersionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access Token（15分有効、メモリ保持）
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF Cookie `notehub_csrf_token` と同値
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VersionId:
      name: versionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Too Many Requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    WorkspaceForbidden:
      description: Workspace access or permission denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ApiResponse:
      type: object
      required: [data]
      properties:
        data: {}
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: 入力値が不正です
            requestId:
              type: string
              format: uuid
...
    SearchUserResponse:
      type: object
      required: [id, email, name, status]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        status:
          type: string
          enum: [active, suspended, deleted]
    Member:
      type: object
      required: [userId, name, email, status, role, joinedAt]
      properties:
        userId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        status:
          type: string
          enum: [active, suspended, deleted]
        role:
          type: string
          enum: [host, member]
        joinedAt:
          type: string
          format: date-time
    AddMemberRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          format: uuid
    AddMemberResponse:
      type: object
      required: [userId, name, email, role, joinedAt]
      properties:
        userId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [member]
        joinedAt:
          type: string
          format: date-time
...
    VersionListItem:
      type: object
      required: [id, versionType, createdBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        versionType:
          type: string
          enum: [auto_save, manual_save, before_restore, restore]
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    GetVersionResponse:
      type: object
      required: [id, documentId, content, versionType, createdBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        documentId:
          type: string
          format: uuid
        content:
          type: string
        versionType:
          type: string
          enum: [auto_save, manual_save, before_restore, restore]
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    SaveManualVersionResponse:
      type: object
      required: [id, createdAt]
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    RestoreVersionResponse:
      type: object
      required: [documentId, versionId, restoredAt]
      properties:
        documentId:
          type: string
          format: uuid
        versionId:
          type: string
          format: uuid
        restoredAt:
          type: string
          format: date-time
...